requirements_persist.txt
absl-py==1.4.0
altair==4.2.2
annotated-types==0.5.0
anyio==3.7.1
astunparse==1.6.3
attrs==23.1.0
beautifulsoup4==4.12.2
blinker==1.6.2
breadability==0.1.20
cachetools==5.3.1
certifi==2023.5.7
cffi==1.16.0
charset-normalizer==3.2.0
click==8.1.5
colorama==0.4.6
decorator==5.1.1
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.3
facedetector-py==0.0.4
fastapi==0.101.1
filelock==3.12.0
filetype==1.2.0
Flask==2.3.2
flatbuffers==23.5.26
fsspec==2023.4.0
gast==0.4.0
gdown==4.7.1
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.54.0
h11==0.14.0
h5py==3.8.0
huggingface-hub==0.14.1
idna==3.4
importlib-metadata==6.8.0
iniconfig==2.0.0
itsdangerous==2.1.2
jax==0.4.8
Jinja2==3.1.2
joblib==1.3.2
jsonschema==4.17.3
keras==2.12.0
lexrank==0.1.0
libclang==16.0.6
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
ml-dtypes==0.1.0
mpmath==1.3.0
networkx==3.1
nltk==3.8.1
numpy==1.23.5
oauthlib==3.2.2
opencv-python==4.8.1.78
openpyxl==3.1.2
opt-einsum==3.3.0
outcome==1.2.0
packaging==23.1
pandas==2.0.3
path.py==12.5.0
Pillow==9.5.0
platformdirs==3.5.0
pluggy==1.3.0
protobuf==4.23.4
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.2.0
pydantic_core==2.6.0
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyrsistent==0.19.3
PySocks==1.7.1
pytest==7.4.2
python-dateutil==2.8.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
regex==2023.8.8
requests==2.31.0
requests-oauthlib==1.3.1
retina-face==0.0.13
rich==13.4.2
rsa==4.9
safetensors==0.3.2
scikit-learn==1.3.0
scipy==1.11.1
selenium==4.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.5
starlette==0.27.0
streamlit==1.24.1
sumy==0.11.0
sympy==1.12
tenacity==8.2.2
tensorboard==2.12.3
tensorboard-data-server==0.7.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-intel==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.1
torchvision==0.15.2
tornado==6.3.2
tqdm==4.66.1
transformers==4.31.0
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
uritools==4.0.1
urlextract==1.8.0
urllib3==2.0.3
validators==0.20.0
watchdog==3.0.0
Werkzeug==2.3.7
wrapt==1.14.1
wsproto==1.2.0
zipp==3.16.2
